+++ src/main/java/net/crypticverse/betterbiomes/block/custom/ThinTreeBlock.java
@@ -28,18 +28,36 @@ package net.crypticverse.betterbiomes.block.custom;
 import com.mojang.serialization.MapCodec;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.AxeItem;
+import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.RotatedPillarBlock;
+import net.minecraft.world.level.block.SimpleWaterloggedBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.block.state.properties.Property;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.level.material.Fluids;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import net.minecraftforge.common.ToolAction;
+import org.jetbrains.annotations.Nullable;
 
-public class ThinTreeBlock extends RotatedPillarBlock {
+public class ThinTreeBlock extends RotatedPillarBlock implements SimpleWaterloggedBlock {
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
+    public static final EnumProperty<Direction.Axis> AXIS = BlockStateProperties.AXIS;
     public static final VoxelShape SHAPE = ThinTreeBlock.box(3, 0, 3, 13, 16, 13);
     public static final MapCodec<ThinTreeBlock> CODEC = simpleCodec(ThinTreeBlock::new);
 
     public ThinTreeBlock(Properties properties) {
         super(properties);
+        this.registerDefaultState((BlockState) this.stateDefinition.any().setValue(WATERLOGGED, false));
     }
 
     @Override
@@ -50,5 +68,23 @@ public class ThinTreeBlock extends RotatedPillarBlock {
     @Override
     public VoxelShape getShape(BlockState p_60555_, BlockGetter p_60556_, BlockPos p_60557_, CollisionContext p_60558_) {
         return SHAPE;
+    }
+
+    @Override
+    public BlockState updateShape(BlockState blockState, Direction direction, BlockState blockState2, LevelAccessor levelAccessor, BlockPos pos, BlockPos pos2) {
+        if ((Boolean) blockState.getValue(WATERLOGGED)) {
+            levelAccessor.scheduleTick(pos, Fluids.WATER, Fluids.WATER.getTickDelay(levelAccessor));
+        }
+        return blockState;
+    }
+
+    @Override
+    public FluidState getFluidState(BlockState blockState) {
+        return (Boolean) blockState.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(blockState);
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(new Property[]{AXIS, WATERLOGGED});
     }
 }